+++ tests/testthat/test-linear-models.R
@@ -113,4 +113,10 @@ if (suppressWarnings(
     ggeffect(fit, c("c12hour", "c172code [high level of education,low level of education]"))
     ggeffect(fit, c("c12hour [exp]", "c172code [high level of education,low level of education]"))
   })
+
+  test_that("ggpredict, lm formula", {
+    ggpredict(fit, ~ c12hour)
+    ggpredict(fit, ~ c12hour + c161sex)
+    ggpredict(fit, ~ c12hour + c161sex + c172code)
+  })
 }