+++ tests/testthat/test-lmer.R
@@ -33,4 +33,4 @@ if (suppressWarnings(
     ggeffect(fit, c("c12hour", "c161sex"))
     ggeffect(fit, c("c12hour", "c161sex", "c172code"))
   })
-^}
+}