+++ NEWS.md
@@ -3,6 +3,7 @@
 ## General
 
 * Reduce package dependencies.
+* Moved package **effects** from dependencies to suggested packages, due to the restrictive requirements (R >= 3.5).
 * New `print()`-method, with a nicer print of the returned data frame. This method replaces the `summary()`-method, which was removed.
 * `ggeffect()` now supports `clm2`-models from the **ordinal**-package.
 
@@ -11,6 +12,8 @@
 * The `type`-argument in `ggpredict()` now has two more options, `type = "fe.zi"` and `type = "re.zi"`, to explicitely condition zero-inflated (mixed) models on their zero-inflation component.
 * `ggpredict()` gets arguments `vcov.fun`, `vcov.type` and `vcov.args` to calculate robust standard errors for confidence intervals of predicted values. These are based on the various `sandwich::vcov*()`-functions, hence robust standard errors can be calculated for all models that are supported by `sandwich::vcov*()`.
 * The `plot()`-method gets two arguments `line.size` and `dot.size`, to determine the size of the geoms.
+* The `ci`-argument for the `plot()`-method now also accepts the character values `"dash"` and `"dot"` to plot dashed or dotted lines as confidence bands.
+* The `terms`-argument in `ggpredict()` and `ggeffect()` may also be a formula, which is more convenient for typing, but less flexible than specifying the terms as character vector with specific options.
 
 ## Bug fixes